Removing FITS wrapper
On principle, any FITS handling should be done by blimpy --
file format should be abstracted away and the blimpy API
should be used only.

Code coverage showed that 0% of fits_wrapper.py was called
in the tests. If we need to add FITS support in the future,
we should enforce that it is done in blimpy.